rename: Class 이름 변경, file 옮김 #50
