+++ src/main/java/com/sounganization/botanify/domain/chat/dto/res/ErrorMessageDto.java
@@ -1,4 +1,4 @@
-package com.sounganization.botanify.domain.chat.dto;
+package com.sounganization.botanify.domain.chat.dto.res;
 
 public record ErrorMessageDto(
         String error
